Complete-Sheet#1

+++ Sheet_1/Coordinates_of_Point.java
@@ -5,8 +5,8 @@ import java.util.Scanner;
 public class Coordinates_of_Point {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double x = sc.nextDouble();
-        double y = sc.nextDouble();
+        long x = sc.nextLong();
+        long y = sc.nextLong();
 
 
         if (x == 0 && y == 0) {
